backend/pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Paweł Rozenkranc <p.rozenkranc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
llama-index = "^0.10.64"
llama-index-llms-ollama = "^0.2.2"
llama-index-readers-web = "^0.1.23"
llama-index-embeddings-ollama = "^0.1.3"
llama-index-vector-stores-qdrant = "^0.2.15"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
